pyproject.toml
[project]
name = "mcp-terminal-client"
version = "0.1.0"
description = "A Python MCP terminal client"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "anyio>=4.9.0",
    "jsonschema>=4.23.0",
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
]
